requirements.txt
asgiref==3.3.1
certifi==2020.12.5
coverage==5.3.1
Django==3.1.4
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.5
sqlparse==0.4.1
